requirements.txt
flask
python-dotenv
azure-cosmos
azure-search-documents
azure-core